refactor: Lighthouse 성능개선 - head에 link로 preconnet 추가 - 페이로드 높은 폰트 woff -> woff2로 변경 - 검색섹션의 svg파일 WebP파일로 변경
