+++ frontend/tracker-app/public/index.html
@@ -2,6 +2,7 @@
 <html lang="en" style="height: 100%">
   <head>
     <meta charset="utf-8" />
+    <link rel="preconnect" href="https://comission-platform.shop" />
     <link rel="apple-touch-icon" sizes="180x180" href="/apple-touch-icon.png" />
     <link rel="icon" type="image/png" sizes="32x32" href="/favicon-32x32.png" />
     <link rel="icon" type="image/png" sizes="16x16" href="/favicon-16x16.png" />
@@ -10,7 +11,7 @@
     <meta name="theme-color" content="#000000" />
     <meta
       name="description"
-      content="Web site created using create-react-app"
+      content="commission platform service. You can create a mission, share it with others, and work together to achieve it."
     />
     <title>Comission Platform</title>
   </head>
